implemented all requirments and updated the output values

+++ src/main/java/com/example/topologyapi/TopologyOperations/TopologyOperationsController.java
@@ -6,6 +6,8 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.io.FileNotFoundException;
 
@@ -30,36 +32,67 @@ public class TopologyOperationsController implements ITopologyOperations {
     @RequestMapping("readJSON")
     @PostMapping
     @Override
-    public void readJSON(@RequestBody String file) throws ParseException, FileNotFoundException {
+    public ResponseEntity<JSONObject> readJSON(@RequestBody String file) throws ParseException, FileNotFoundException {
+        JSONObject message = new JSONObject();
         JSONObject object = (JSONObject) new JSONParser().parse(file);
+        if(object.get("fileName")==null){
+            message.put("message", "Invalid Input");
+            return new ResponseEntity<JSONObject>(message, HttpStatus.BAD_REQUEST);
+        }
         String fileName = object.get("fileName").toString();
-        database.readJSON(fileName);
+        return database.readJSON(fileName);
     }
 
     @RequestMapping("writeJSON")
     @PostMapping
     @Override
-    public void writeJSON(@RequestBody String topologyId) throws ParseException {
+    public ResponseEntity<JSONObject> writeJSON(@RequestBody String topologyId) throws ParseException {
         JSONObject object = (JSONObject) new JSONParser().parse(topologyId);
+        JSONObject message = new JSONObject();
+        if(object.get("topologyId")==null){
+            message.put("message", "Invalid Input");
+            return new ResponseEntity<JSONObject>(message, HttpStatus.BAD_REQUEST);
+        }
         String topologyID = object.get("topologyId").toString();
-        database.writeJSON(topologyID);
+        return database.writeJSON(topologyID);
     }
 
     @RequestMapping("deleteTopology")
     @PostMapping
     @Override
-    public void deleteTopology(@RequestBody String topologyId) throws ParseException {
+    public ResponseEntity<JSONObject> deleteTopology(@RequestBody String topologyId) throws ParseException {
         JSONObject object = (JSONObject) new JSONParser().parse(topologyId);
+        JSONObject message = new JSONObject();
+        if(object.get("topologyId") == null){
+            message.put("message", "Invalid Input");
+            return new ResponseEntity<JSONObject>(message, HttpStatus.BAD_REQUEST);
+        }
         String topologyID = object.get("topologyId").toString();
-        database.deleteTopology(topologyID);
+        return database.deleteTopology(topologyID);
     }
 
     @RequestMapping("queryDevices")
     @PostMapping
     @Override
-    public JSONArray queryDevices(@RequestBody String topologyId) throws ParseException {
+    public ResponseEntity<JSONArray> queryDevices(@RequestBody String topologyId) throws ParseException {
         JSONObject object = (JSONObject) new JSONParser().parse(topologyId);
+        if(object.get("topologyId") == null){
+            return new ResponseEntity<JSONArray>(new JSONArray(), HttpStatus.BAD_REQUEST);
+        }
         String topologyID = object.get("topologyId").toString();
         return database.queryDevices(topologyID);
+    }
+
+    @RequestMapping("queryDevicesToNetlist")
+    @PostMapping
+    @Override
+    public ResponseEntity<JSONArray> queryDevicesToNetlist(@RequestBody String ids) throws ParseException {
+        JSONObject object = (JSONObject) new JSONParser().parse(ids);
+        if(object.get("topologyId") == null || object.get("netlistId") == null){
+            return new ResponseEntity<JSONArray>(new JSONArray(), HttpStatus.BAD_REQUEST);
+        }
+        String topologyID = object.get("topologyId").toString();
+        String netlistId = object.get("netlistId").toString();
+        return database.queryDevicesToNetlist(topologyID, netlistId);
     }
 }
